+++ lib/translators/DoodadsTranslator.ts
@@ -28,8 +28,8 @@ enum flag {
 }
 
 export class DoodadsTranslator {
-    public _outBufferToWar: HexBuffer;
-    public _outBufferToJSON: W3Buffer;
+    private _outBufferToWar: HexBuffer;
+    private _outBufferToJSON: W3Buffer;
 
     constructor() { }