+++ lib/translators/InfoTranslator.ts
@@ -111,8 +111,8 @@ interface Force {
 
 export class InfoTranslator {
 
-    public _outBufferToWar: HexBuffer;
-    public _outBufferToJSON: W3Buffer;
+    private _outBufferToWar: HexBuffer;
+    private _outBufferToJSON: W3Buffer;
 
     constructor() { }