+++ lib/translators/RegionsTranslator.ts
@@ -19,8 +19,8 @@ interface Rect {
 
 export class RegionsTranslator {
 
-    public _outBufferToWar: HexBuffer;
-    public _outBufferToJSON: W3Buffer;
+    private _outBufferToWar: HexBuffer;
+    private _outBufferToJSON: W3Buffer;
 
     constructor() { }