+++ lib/translators/TerrainTranslator.ts
@@ -23,8 +23,8 @@ interface Offset {
 
 export class TerrainTranslator {
 
-    public _outBufferToWar: HexBuffer;
-    public _outBufferToJSON: W3Buffer;
+    private _outBufferToWar: HexBuffer;
+    private _outBufferToJSON: W3Buffer;
 
     constructor() { }