+++ lib/translators/UnitsTranslator.ts
@@ -39,8 +39,8 @@ interface Abilities {
 
 export class UnitsTranslator {
 
-    public _outBufferToWar: HexBuffer;
-    public _outBufferToJSON: W3Buffer;
+    private _outBufferToWar: HexBuffer;
+    private _outBufferToJSON: W3Buffer;
 
     constructor() { }