+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## FEATURES
 ## FIXES
  * Resolve `[DEP0005] DeprecationWarning: Buffer()` warning in `HexBuffer.ts`
+ * Fix scoping issues on `*Translator.ts`, `HexBuffer.ts` and `W3Buffer.ts` where certain fields that should be `private` were marked as `public`
 ## MAINTENANCE
  * Upgrade to Node 14.x LTS
  * Upgrade to npm 7.x